// backend/src/api/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, KvCache, Logger, MRNObject, Queue, SmartBucket, SmartMemory, SqlDatabase, Tracer, } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CEREBRAS_API_KEY: string;
  CEREBRAS_API_URL: string;
  EXTENSION_DB: SqlDatabase;
  EXTENSION_STORAGE: SmartBucket;
  FRONTEND_URL: string;
  GENERATION_QUEUE: Queue<import('../job-processor').Body>;
  logger: Logger;
  mem: KvCache;
  RATE_LIMITER: KvCache;
  STRIPE_PRICE_ID: string;
  STRIPE_SECRET_KEY: string;
  STRIPE_WEBHOOK_SECRET: string;
  tracer: Tracer;
  USER_CONTEXT: SmartMemory;
  WORKOS_API_KEY: string;
  WORKOS_CLIENT_ID: string;
  WORKOS_COOKIE_PASSWORD: string;
  WORKOS_REDIRECT_URI: string;
}